Properly update header for non-animated transitions and on old i… (#384)
This change fixes an issue when navigation header wasn't properly updated on mount time. The issue was due to us trying to animate transition while adding initial set of push VCs should never be animated. On top of that we observed that willShow delegate method wasn't getting triggered in some cases on old iOS devices. We added additional call to updating the header during mount time to workaround that.