Introduce custom remove action to cleanup dismissed screens. (#390)
This change introduces REMOVE custom navigation command to be used for navigation wrapper. This command works similarily to POP however it does not pop views on top of the one passed as a key. It is important that when view is dismissed we only remove that screen from the stack and not accidently remove other views that might've been added on top in the meantime.