fix: trigger color change on every update (#552)
It is a workaround for loading custom back icon when transitioning from a screen without header to the screen which has one. This action fails when navigating to the screen with the header for the second time and loads the default back button. It looks like changing the tint color of the navbar triggers an update of the items belonging to it and it seems to load the custom back image so we change the tint color's alpha by a very small amount and then set it to the one it should have.