[native-stack] Add transitionStart and transitionEnd events (#499)
Implement transitionStart and transitionEnd events in native stack, similar to the events available in js stack. On Android there isn't really a concept of nested stacks and it uses only one fragment manager. This means a single "transition" so we need a way to trigger this transition on all nested stacks of a screen. To do that we added a way for nested stacks to register on their parent screen so when a screen is dismissed it can also call the dismiss method on all its child stacks.