feat: add slide animations on Android (#648)
Added `slide_from_left` and `slide_from_right` to the list of available transition animations on Android. They are mapped to default transition on iOS. Also, the ordering of the screens adding/removing in `ScreenStack.java`  was changed because the custom animations need to be set first before any screens are added/removed.