feat: add setting frame in notifyWillAppear (#772)
Suggested by @kmagiera change adding setting frame in the `viewWillAppear`. It ensures that the `parentViewController` is set when trying to set the frame, and if the parent is not `UINavigationController`, then the frame is properly set. It fixes the issue that appeared after #712 was merged, where there is a check for `parentViewController` not being nil introduced.