feat: rewrite v4 native-stack to TS (#793)
PR migrating v4 native-stack to TS.

Many options available in v4 native-stack are taken from stack-navigator, they are included in StackNavigatorOptions and NativeStackNavigationConfig types. There is also a type that is not used in v5 native-stack and gives an option to use HeaderBackButton from react-navigation-stack, it is called BackButtonProps.

Some of the props used for HeaderBackButton have been changed because the former props didn't match the actual props from HeaderBackButton. We keep the former props for backwards compatibility. Changed props are:

backButtonTitle: resolves to label instead of title,
truncatedBackButtonTitle: resolves to truncatedLabel instead of truncatedTitle,
backTitleVisible: resolves to labelVisible instead of backTitleVisible,
headerBackTitleStyle: resolves to labelStyle instead of titleStyle,
layoutPreset: resolves to titleLayout instead of layoutPreset,
scene was removed