Add missing dependency on React-RCTImage pod. (#934)
When building for iOS with use_frameworks! in the project Podfile Xcode became fussy about explicitly specifying the framework dependencies rather than just the "React" target. In general RN modules have changed to depend on "React-Core" and all is fine. However, modules that use classes that aren't in the "React-Core" pod need to name all their dependencies explicitly. I've checked and every class that is currently imported by the native iOS code comes from "React-Core" except for RCTImageView and RCTImageLoader. These are both in the "React-RCTImage" framework/pod. Added this to the dependencies to fix linker errors for some projects that use this module.