feat: add status bar manipulation on Android (#863)
Added status bar manipulation on Android as addition to #860. It takes the native code provided by react-native and applies react-native-screens logic of resolving the proper Screen to change the behavior. Also changed the logic of taking the proper screen into consideration on iOS.

These changes are done to make it available to control the status bar and orientation in all navigators of react-navigation, which use Screen component.

- Moved all props regarding status bar and orientation management to Screen component.
- Added WindowTraits class on both Android and iOS to manage the static methods used for changing the traits.
- Updated documentation about the props for managing traits.

Co-authored-by: kacperkapusciak <kapusciak.k@gmail.com>