fix: make stack animations work same on both platforms (#1190)
PR properly resolving if there should be an animation of transition between screens on both platforms. The logic is that if we perform push transition, we ask the screen we are transitioning to, and if we do pop transition, we ask the disappearing screen. It works the same for replace animation, so if the screen we are transitioning to specifies replaceAnimation: 'push', we push it and ask it for animation, and if it specifies replaceAnimation: 'pop', we ask the screen which is being replaced.